requirements.txt
openai
pydantic
boto3
python-dotenv
requests
moviepy
pysrt
Pillow
numpy
srt